implement settings provider 'custom' implementation of taffy tree to allow for direct usage of CssStyle change color to use four u8s instead of four f32s change a bunch of Strings to Arc<str>s
need to fix - some styles not working properly (specifically beatmap select map buttons not in a column for some reason), need to fix some color things (specifically some storyboard elements being colored improperly)

+++ crates/tataku-engine/src/graphics/drawables/osu_slider.rs
@@ -5,7 +5,7 @@ pub struct SliderDrawable {
     /// bounding size of the slider
     pub size: Vector2,
     /// alpha of whole drawable, mainly used for fade in
-    pub alpha: f32,
+    pub alpha: u8,
 
 
     /// slider data to be passed onto the shader
